# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for the infrastructure archiving system
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: getDocuments
      tags: [documents]
      summary: List documents
      parameters:
        - name: id_projet
          in: query
          schema:
            type: integer
        - name: id_lot
          in: query
          schema:
            type: integer
        - name: id_phase
          in: query
          schema:
            type: integer
        - name: statut
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a document (metadata only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /documents/{id}/upload:
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a file for a document (creates new version)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                commentaire:
                  type: string
      responses:
        "200":
          description: Upload successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Version"
  /documents/{id}/archiver:
    post:
      operationId: archiverDocument
      tags: [documents]
      summary: Archive a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                commentaire:
                  type: string
      responses:
        "200":
          description: Document archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{id}/versions:
    get:
      operationId: getDocumentVersions
      tags: [documents]
      summary: Get all versions of a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Version"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Document:
      type: object
      properties:
        id_document:
          type: integer
        nom_doc:
          type: string
        chemin:
          type: ["string", "null"]
        is_global:
          type: boolean
        date_creation:
          type: ["string", "null"]
        id_phase:
          type: ["integer", "null"]
        commentaire:
          type: ["string", "null"]
        id_version:
          type: ["integer", "null"]
        id_projet:
          type: ["integer", "null"]
        id_type:
          type: ["integer", "null"]
        statut:
          type: ["string", "null"]
        id_utilisateur:
          type: ["integer", "null"]
        nom_projet:
          type: ["string", "null"]
        nom_phase:
          type: ["string", "null"]
        lib_type:
          type: ["string", "null"]
        nom_auteur:
          type: ["string", "null"]
        numero_version:
          type: ["integer", "null"]
      required:
        - id_document
        - nom_doc
    CreateDocumentBody:
      type: object
      properties:
        nom_doc:
          type: string
        is_global:
          type: boolean
        id_phase:
          type: integer
        commentaire:
          type: string
        id_projet:
          type: integer
        id_type:
          type: integer
        id_utilisateur:
          type: integer
      required:
        - nom_doc
    UpdateDocumentBody:
      type: object
      properties:
        nom_doc:
          type: string
        is_global:
          type: boolean
        id_phase:
          type: integer
        commentaire:
          type: string
        id_projet:
          type: integer
        id_type:
          type: integer
        statut:
          type: string
    Version:
      type: object
      properties:
        id_version:
          type: integer
        numero_version:
          type: integer
        date_modification:
          type: ["string", "null"]
        fichier_path:
          type: ["string", "null"]
        commentaire:
          type: ["string", "null"]
        id_document:
          type: ["integer", "null"]
      required:
        - id_version
        - numero_version
...